refactor: use DashboardToolsetDateDropdown rather than DashboardToolset
