+++ apps/web/src/services/dashboards/components/DashboardCustomize.vue
@@ -16,11 +16,12 @@ import DashboardCustomizePageName
 import DashboardCustomizeSidebar from '@/services/dashboards/components/DashboardCustomizeSidebar.vue';
 import DashboardLabels from '@/services/dashboards/components/DashboardLabels.vue';
 import DashboardRefreshDropdown from '@/services/dashboards/components/DashboardRefreshDropdown.vue';
-import DashboardToolset from '@/services/dashboards/components/DashboardToolset.vue';
+import DashboardToolsetDateDropdown from '@/services/dashboards/components/DashboardToolsetDateDropdown.vue';
 import DashboardVariables from '@/services/dashboards/components/DashboardVariables.vue';
 import DashboardWidgetContainer from '@/services/dashboards/components/DashboardWidgetContainer.vue';
 import { DASHBOARDS_ROUTE } from '@/services/dashboards/routes/route-constant';
 import { useDashboardDetailInfoStore } from '@/services/dashboards/stores/dashboard-detail-info-store';
+
 
 interface Props {
     dashboardId?: string;
@@ -92,7 +93,9 @@ onBeforeUnmount(() => {
         />
         <div class="filters-box">
             <dashboard-labels editable />
-            <dashboard-toolset />
+            <dashboard-toolset-date-dropdown v-show="dashboardDetailState.settings.date_range.enabled"
+                                             :date-range="dashboardDetailState.settings.date_range"
+            />
         </div>
         <p-divider />
         <div class="dashboard-selectors">
